Add unit test for .po files
#267
