+++ test/locale.spec.ts
@@ -1,0 +1,8 @@
+import { loadTranslations } from "../src/locale";
+
+describe(".po files", () => {
+	test("syntax valid", () => {
+		const locales = loadTranslations();
+		expect(locales.length).toBeGreaterThan(0);
+	});
+});
